refactor: move claimer from strategy to account module
