+++ src/account/mod.rs
@@ -1,0 +1,4 @@
+//! Account-level operations (claiming, fund management, etc.)
+
+#[cfg(feature = "claimer_daemon")]
+pub mod claimer;
